# Optimize is_min_level() function