This adds a couple of simple push/pop_min/pop_max benchmarks and applies a minor optimization to the `is_min_level()` function. Benchmarks before/after:

```
Before:
test pop_max_seq ... bench:      53,099 ns/iter (+/- 2,666)
test pop_min_seq ... bench:      77,194 ns/iter (+/- 3,006)
test push_seq    ... bench:      13,041 ns/iter (+/- 326)

After:
test pop_max_seq ... bench:      50,055 ns/iter (+/- 2,839)
test pop_min_seq ... bench:      74,374 ns/iter (+/- 3,805)
test push_seq    ... bench:      10,139 ns/iter (+/- 237)
```