@tov 2018-05-03
That's great, thanks! I knew there had to be a better way to do that.

Are you using this crate for something? I just made it for fun and didn’t know anyone had even noticed it.