@nikic 2018-05-04
I'm using this crate to implement the stack algorithm (sequential decoding). While it's called stack algorithm, the structure it logically uses is a finite-size heap, which can be implemented using a min-max-heap. Unfortunately heaps are pretty slow, so I'll probably have to move to some kind of binning scheme for better performance, but using a min-max-heap makes the initial implementation straightforward.